Move redis, sshd and tomcat tasks into analysis folder (#1425)
