+++ turbinia/jobs/tomcat.py
@@ -15,7 +15,7 @@
 """Job to execute Apache Tomcat analysis task."""
 from __future__ import unicode_literals
 from turbinia.workers import artifact
-from turbinia.workers import tomcat
+from turbinia.workers.analysis import tomcat
 from turbinia.evidence import ContainerdContainer
 from turbinia.evidence import Directory
 from turbinia.evidence import DockerContainer
